organizing event log clients
