+++ event-log-verification-service/src/main/scala/com/ubirch/verification/LookupServiceBinder.scala
@@ -12,7 +12,7 @@ import com.ubirch.services.config.ConfigProvider
 import com.ubirch.services.execution.ExecutionProvider
 import com.ubirch.services.lifeCycle.{ DefaultJVMHook, DefaultLifecycle, JVMHook, Lifecycle }
 import com.ubirch.verification.services._
-import com.ubirch.verification.services.eventlog.{ CachedEventLogClient, EventLogClient, NewEventLogClient }
+import com.ubirch.verification.services.eventlog.{ CachedEventLogClient, DefaultEventLogClient, EventLogClient }
 import com.ubirch.verification.util.udash.JettyServer
 
 import scala.concurrent.ExecutionContext
@@ -20,34 +20,17 @@ import scala.concurrent.ExecutionContext
 class LookupServiceBinder extends AbstractModule with BasicServices with CassandraServices with ExecutionServices {
 
   def lifecycle: ScopedBindingBuilder = bind(classOf[Lifecycle]).to(classOf[DefaultLifecycle])
-
   def jvmHook: ScopedBindingBuilder = bind(classOf[JVMHook]).to(classOf[DefaultJVMHook])
-
   def config: ScopedBindingBuilder = bind(classOf[Config]).toProvider(classOf[ConfigProvider])
-
   def executionContext: ScopedBindingBuilder = bind(classOf[ExecutionContext]).toProvider(classOf[ExecutionProvider])
-
-  //Cassandra Cluster
   def clusterService: ScopedBindingBuilder = bind(classOf[ClusterService]).to(classOf[DefaultClusterService])
-
   def connectionService: ScopedBindingBuilder = bind(classOf[ConnectionService]).to(classOf[DefaultConnectionService])
-
-  //Cassandra Cluster
-
   def finder: ScopedBindingBuilder = bind(classOf[Finder]).to(classOf[DefaultFinder])
-
   def gremlin: ScopedBindingBuilder = bind(classOf[Gremlin]).to(classOf[DefaultGremlinConnector])
-
-  //CachedEventLogClient
-
-  def newEventLogClient: ScopedBindingBuilder = bind(classOf[EventLogClient]).annotatedWith(Names.named("New")).to(classOf[NewEventLogClient])
-
+  def eventLogClient: ScopedBindingBuilder = bind(classOf[EventLogClient]).to(classOf[DefaultEventLogClient])
   def cachedEventLogClient: ScopedBindingBuilder = bind(classOf[EventLogClient]).annotatedWith(Names.named("Cached")).to(classOf[CachedEventLogClient])
-
   def redisOpt: ScopedBindingBuilder = bind(classOf[RedisCache]).toProvider(classOf[RedisProvider])
-
   def healthCheck: ScopedBindingBuilder = bind(classOf[HealthCheckServer]).toProvider(classOf[HealthCheckProvider])
-
   def jettyServer: ScopedBindingBuilder = bind(classOf[JettyServer]).toProvider(classOf[JettyServerProvider])
 
   override def configure(): Unit = {
@@ -59,7 +42,7 @@ class LookupServiceBinder extends AbstractModule with BasicServices with Cassand
     executionContext
     gremlin
     finder
-    newEventLogClient
+    eventLogClient
     cachedEventLogClient
     redisOpt
     healthCheck
