+++ event-log-verification-service/src/main/scala/com/ubirch/verification/models/LookupResult.scala
@@ -3,20 +3,20 @@ package com.ubirch.verification.models
 import org.json4s.JValue
 import org.json4s.JsonAST.JNull
 
-case class LookupResult(success: Boolean, key: String, queryType: String, message: String, event: JValue, anchors: JValue)
+case class LookupResult(success: Boolean, value: String, queryType: String, message: String, event: JValue, anchors: JValue)
 
 object LookupResult {
 
   def apply(success: Boolean, key: String, queryType: QueryType, message: String, event: JValue, anchors: JValue): LookupResult =
     LookupResult(success, key, queryType.value, message, event, anchors)
 
-  def NotFound(key: String, queryType: QueryType): LookupResult =
-    LookupResult(success = true, key, queryType.value, "Nothing Found", JNull, JNull)
+  def NotFound(value: String, queryType: QueryType): LookupResult =
+    LookupResult(success = true, value, queryType.value, "Nothing Found", JNull, JNull)
 
-  def Found(key: String, queryType: QueryType, event: JValue, anchors: JValue): LookupResult =
-    LookupResult(success = true, key, queryType.value, "Query Successfully Processed", event, anchors)
+  def Found(value: String, queryType: QueryType, event: JValue, anchors: JValue): LookupResult =
+    LookupResult(success = true, value, queryType.value, "Query Successfully Processed", event, anchors)
 
-  def Error(key: String, queryType: QueryType, message: String): LookupResult =
-    LookupResult(success = false, key, queryType.value, message, JNull, JNull)
+  def Error(value: String, queryType: QueryType, message: String): LookupResult =
+    LookupResult(success = false, value, queryType.value, message, JNull, JNull)
 
 }
