+++ event-log-verification-service/src/main/scala/com/ubirch/verification/services/eventlog/CachedEventLogClient.scala
@@ -8,7 +8,7 @@ import javax.inject._
 import scala.concurrent.{ ExecutionContext, Future }
 
 @Singleton
-class CachedEventLogClient @Inject() (@Named("New") underlying: EventLogClient, redis: RedisCache)(implicit ec: ExecutionContext) extends EventLogClient with StrictLogging {
+class CachedEventLogClient @Inject() (underlying: EventLogClient, redis: RedisCache)(implicit ec: ExecutionContext) extends EventLogClient with StrictLogging {
 
   private val getEventByHashCached =
     redis
