+++ event-log-verification-service/src/main/scala/com/ubirch/verification/services/eventlog/EventLogClient.scala
@@ -19,4 +19,5 @@ trait EventLogClient {
       responseForm: ResponseForm,
       blockchainInfo: BlockchainInfo
   ): Future[LookupResult]
+
 }
